src/main/java/com/pontointeligente/utils/PasswordsUtils.java
package com.pontointeligente.utils;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;

public class PasswordsUtils {

    private static final Logger log = LoggerFactory.getLogger(PasswordsUtils.class);

    public PasswordsUtils() {
    }

    /**
     * Gera um hash utlizando o BCrypt
     *
     * @param senha
     * @return String
     */
    public static String gerarBCrypt(String senha) {
        if (senha == null) {
            return senha;
        }

        log.info("Gerando  hash como o BCrypt.");
        BCryptPasswordEncoder bCryptEncoder = new BCryptPasswordEncoder();
        return bCryptEncoder.encode(senha);
    }
}
